addressed izzys comments, added cls as a kwarg
