+++ sdk/tables/azure-data-tables/azure/data/tables/aio/_base_client_async.py
@@ -31,7 +31,7 @@ from .._policies import (
     StorageHeadersPolicy
 )
 from ._policies_async import AsyncStorageResponseHook
-from .._deserialize import _process_table_error
+from .._error import _process_table_error
 from .._models import PartialBatchErrorException
 
 if TYPE_CHECKING:
